requirements.txt
Django==3.0.6
djangorestframework==3.11.0
secrets==1.0.2
